+++ pcs/lib/commands/resource.py
@@ -361,6 +361,7 @@ def create(
     ensure_disabled: bool = False,
     wait: WaitType = False,
     allow_not_suitable_command: bool = False,
+    enable_agent_self_validation: bool = False,
 ):
     # pylint: disable=too-many-arguments, too-many-locals
     """
@@ -394,6 +395,8 @@ def create(
         pcs.lib.commands.remote_node);
         in the case of remote/guest node forcible error is produced when this
         flag is set to False and warning is produced otherwise
+    enable_agent_self_validation -- if True, use agent self-validation feature
+        to validate instance attributes
     """
     runner = env.cmd_runner()
     agent_factory = ResourceAgentFacadeFactory(runner, env.report_processor)
@@ -440,6 +443,7 @@ def create(
             allow_invalid_operation,
             allow_invalid_instance_attributes,
             use_default_operations,
+            enable_agent_self_validation=enable_agent_self_validation,
         )
         if env.report_processor.has_errors:
             raise LibraryError()
@@ -465,6 +469,7 @@ def create_as_clone(
     wait: WaitType = False,
     allow_not_suitable_command: bool = False,
     allow_incompatible_clone_meta_attributes: bool = False,
+    enable_agent_self_validation: bool = False,
 ):
     # pylint: disable=too-many-arguments, too-many-locals
     """
@@ -495,6 +500,8 @@ def create_as_clone(
     allow_incompatible_clone_meta_attributes -- if True some incompatible clone
         meta attributes are treated as a warning, or as a forceable error if
         False
+    enable_agent_self_validation -- if True, use agent self-validation feature
+        to validate instance attributes
     """
     runner = env.cmd_runner()
     agent_factory = ResourceAgentFacadeFactory(runner, env.report_processor)
@@ -579,6 +586,7 @@ def create_as_clone(
             allow_invalid_operation,
             allow_invalid_instance_attributes,
             use_default_operations,
+            enable_agent_self_validation=enable_agent_self_validation,
         )
 
         clone_element = resource.clone.append_new(
@@ -609,6 +617,7 @@ def create_in_group(
     put_after_adjacent: bool = False,
     wait: WaitType = False,
     allow_not_suitable_command: bool = False,
+    enable_agent_self_validation: bool = False,
 ):
     # pylint: disable=too-many-arguments, too-many-locals
     """
@@ -637,6 +646,8 @@ def create_in_group(
         adjacent resource
     wait -- is flag for controlling waiting for pacemaker idle mechanism
     allow_not_suitable_command -- turn forceable errors into warnings
+    enable_agent_self_validation -- if True, use agent self-validation feature
+        to validate instance attributes
     """
     runner = env.cmd_runner()
     agent_factory = ResourceAgentFacadeFactory(runner, env.report_processor)
@@ -712,6 +723,7 @@ def create_in_group(
             allow_invalid_operation,
             allow_invalid_instance_attributes,
             use_default_operations,
+            enable_agent_self_validation=enable_agent_self_validation,
         )
         if ensure_disabled:
             resource.common.disable(primitive_element, id_provider)
@@ -749,6 +761,7 @@ def create_into_bundle(
     wait: WaitType = False,
     allow_not_suitable_command: bool = False,
     allow_not_accessible_resource: bool = False,
+    enable_agent_self_validation: bool = False,
 ):
     # pylint: disable=too-many-arguments, too-many-locals
     """
@@ -776,6 +789,8 @@ def create_into_bundle(
     wait -- is flag for controlling waiting for pacemaker idle mechanism
     allow_not_suitable_command -- turn forceable errors into warnings
     allow_not_accessible_resource -- turn forceable errors into warnings
+    enable_agent_self_validation -- if True, use agent self-validation feature
+        to validate instance attributes
     """
     runner = env.cmd_runner()
     agent_factory = ResourceAgentFacadeFactory(runner, env.report_processor)
@@ -823,6 +838,7 @@ def create_into_bundle(
             allow_invalid_operation,
             allow_invalid_instance_attributes,
             use_default_operations,
+            enable_agent_self_validation=enable_agent_self_validation,
         )
         if ensure_disabled:
             resource.common.disable(primitive_element, id_provider)