+++ pcs/resource.py
@@ -624,6 +624,7 @@ def _format_desc(indentation, desc):
 def resource_create(lib, argv, modifiers):
     """
     Options:
+      * --agent-validation - use agent self validation of instance attributes
       * --before - specified resource inside a group before which new resource
         will be placed inside the group
       * --after - specified resource inside a group after which new resource
@@ -637,6 +638,7 @@ def resource_create(lib, argv, modifiers):
       * -f - CIB file
     """
     modifiers.ensure_only_supported(
+        "--agent-validation",
         "--before",
         "--after",
         "--group",
@@ -699,6 +701,7 @@ def resource_create(lib, argv, modifiers):
         use_default_operations=not modifiers.get("--no-default-ops"),
         wait=modifiers.get("--wait"),
         allow_not_suitable_command=modifiers.get("--force"),
+        enable_agent_self_validation=modifiers.get("--agent-validation"),
     )
 
     clone_id = parts.get("clone_id", None)
@@ -966,6 +969,7 @@ def update_cmd(lib: Any, argv: List[str], modifiers: InputModifiers) -> None:
     """
     Options:
       * -f - CIB file
+      * --agent-validation - use agent self validation of instance attributes
       * --wait
       * --force - allow invalid options, do not fail if not possible to get
         agent metadata, allow not suitable command
@@ -982,11 +986,14 @@ def resource_update(args: List[str], modifiers: InputModifiers) -> None:
     """
     Commandline options:
       * -f - CIB file
+      * --agent-validation - use agent self validation of instance attributes
       * --wait
       * --force - allow invalid options, do not fail if not possible to get
         agent metadata, allow not suitable command
     """
-    modifiers.ensure_only_supported("-f", "--wait", "--force")
+    modifiers.ensure_only_supported(
+        "-f", "--wait", "--force", "--agent-validation"
+    )
     if len(args) < 2:
         raise CmdLineInputError()
     res_id = args.pop(0)
@@ -1052,6 +1059,9 @@ def resource_update(args: List[str], modifiers: InputModifiers) -> None:
             res_id,
             get_resources(lib_pacemaker.get_cib(cib_xml)),
             force=bool(modifiers.get("--force")),
+            enable_agent_self_validation=bool(
+                modifiers.get("--agent-validation")
+            ),
         )
         if report_list:
             process_library_reports(report_list)