+++ pcs_test/tier1/cib_resource/test_stonith_create.py
@@ -45,7 +45,6 @@ class PlainStonith(ResourceTest):
                     </operations>
                 </primitive>
             </resources>""",
-            stderr_start="Warning: Validation result from agent:",
         )
 
     def test_error_when_not_valid_name(self):
@@ -251,7 +250,6 @@ class WithMeta(ResourceTest):
                     </operations>
                 </primitive>
             </resources>""",
-            stderr_start="Warning: Validation result from agent:",
         )